engine-prototype/requirements.txt
# engine-prototype/requirements.txt
PyYAML>=6.0
semgrep>=1.0.0
bandit>=1.7.0
safety>=2.0.0
xmltodict>=0.13.0
rich>=13.0.0

# Para github_analyzer.py (si lo mantenemos)
gitpython>=3.1.0
requests>=2.31.0
click>=8.1.0
pygments>=2.15.0